requirements.txt
GitPython==3.1.43
matplotlib==3.8.2